requirements.txt
pillow
pyperclip
pyimgur